+++ main.js
@@ -8,8 +8,7 @@
 // console.log(records);
 
 import '/main.scss';
-import '@/layout/header';
-import '@/layout/footer';
+import '@/layout/index';
 import 'swiper/css/bundle';
 import Swiper from 'swiper/bundle';
