--- src/pages/header/index.html
@@ -1,17 +1,0 @@
-<!doctype html>
-<html lang="ko-KR">
-  <head>
-    <title>Vite Scss</title>
-    <meta charset="UTF-8" />
-    <link rel="icon" type="image/svg+xml" href="/vite.svg" />
-    <meta name="viewport" content="width=device-width, initial-scale=1.0" />
-    <meta
-      http-equiv="Content-Security-Policy"
-      content="upgrade-insecure-requests"
-    />
-    <script type="module" src="/src/pages/header/header.js"></script>
-  </head>
-  <body>
-    <c-header></c-header>
-  </body>
-</html>
