+++ src/styles/layout/header.scss
@@ -1,38 +1,39 @@
 @use '@/styles/index' as *;
 
 body {
-  min-width: 320px;
-  width: 100%;
-  background-color: $brand-dark-1;
   position: relative;
+  font-size: $text-12;
 }
 
 .header {
+  background-color: red;
   width: 100%;
   height: 38px;
   padding: 10px 12px;
   position: absolute;
+  top: 0;
+  left: 0;
 
   .nav {
-    @include flexbox(row, nowrap, space-between, center, null);
+    @include flexbox(row, nowrap, space-between, center, $spacing-8);
     color: $gray-200;
   }
-  
+
   .header__logo {
-    display: inline-block;
-    margin-right: $spacing-28;
+    width: 46px;
     height: 15px;
-  
+
     .header__logo__link {
-      display: inline-block;
+      width: 100%;
       height: 15px;
 
       .logo1 {
+        width: 100%;
         height: 15px;
       }
     }
   }
-  
+
   .header__menu {
     @include flexbox(row, nowrap, null, center, $spacing-28);
     flex-grow: 1;
@@ -41,28 +42,32 @@ body {
 
   .menu__list {
     font-size: $text-12;
-    line-height: $line-160;
   }
 
   .list__live {
-    @include flexbox(row, nowrap, center, center, 4px);
+    @include flexbox(row, nowrap, center, center, $spacing-4);
     width: 56px;
-    
-    img {
+
+    svg {
       width: 20px;
+      height: 15px;
     }
   }
 
-  .list__paramount{
+  .list__paramount {
     width: 65px;
+    height: 15px;
 
-    img {
+    svg {
       width: 65px;
+      height: 15px;
     }
   }
 
-  .header__actions{
+  .header__actions {
     @include flexbox(row, nowrap, center, center, $spacing-16);
+    width: 50px;
+    height: 15px;
 
     .actions__search {
       width: 18px;
@@ -74,16 +79,108 @@ body {
   }
 }
 
+// 프로필 버튼
+.button_profile_open {
+  display: none;
+  width: 18px;
+  height: 18px;
+}
+
+.profile_container {
+  img {
+    width: 114px;
+    height: 114px;
+  }
+}
+
+.profile {
+  display: none;
+  position: fixed;
+  left: 0;
+  top: 0;
+  translate: (-50%, -50%);
+  background: rgba(0, 0, 0, 0.6);
+
+  width: 100%;
+  height: 100%;
+}
+
+.profile_wrapper {
+  position: absolute;
+  left: 50%;
+  top: 50%;
+  transform: translate(-50%, -50%);
+  background-color: $brand-dark-1;
+  width: 20rem;
+  margin: 5 auto;
+}
+
+.active {
+  display: block;
+}
+
+.header {
+  .header-signin {
+    display: flex;
+  }
+}
+
+.header-signin {
+  display: block;
+}
+
+.profile_container {
+  @include flexbox(column, nowrap, null, center, $spacing-12);
+
+  margin: 0 0 $spacing-32;
+  font-size: $text-16;
+
+  h2 {
+    font-size: $text-16;
+    color: $gray-400;
+  }
+  button {
+    @include size(18rem, 2.625rem);
+    background-color: $brand-dark-1;
+    font-size: $text-12;
+    color: $gray-500;
+    border: 1px solid $gray-500;
+    border-radius: 0.25rem;
+  }
+}
+
+.button_container {
+  @include flexbox(column, nowrap, null, center, $spacing-16);
+
+  button {
+    @include size(18rem, 2.625rem);
+    padding: 0 $spacing-12;
+    font-size: $text-12;
+    background-color: $brand-dark-2;
+    color: $gray-500;
+
+    img {
+      @include size($spacing-16, $spacing-16);
+    }
+  }
+
+  .deleteID_button {
+    background-color: $brand-dark-1;
+  }
+}
+
+// 미디어 쿼리
+
 @include mq(tablet) {
   .header {
     height: 56px;
     padding: 16px 40px;
-    
+
     .header__logo {
       display: inline-block;
       margin-right: $spacing-28;
-      height:  24px;
-    
+      height: 24px;
+
       .header__logo__link {
         display: inline-block;
         height: 24px;
@@ -93,12 +190,12 @@ body {
         }
       }
     }
-    
+
     .header__menu {
-      display: flex;
+      // display: flex;
     }
 
-    .header__actions{
+    .header__actions {
       @include flexbox(row, nowrap, center, center, $spacing-24);
 
       .actions__search {
@@ -112,61 +209,65 @@ body {
   }
 }
 
-@include mq(desktop) {
-  .header {
-    height: 100px;
-    padding: 29px 70px;
-    background: linear-gradient(to bottom, rgba($brand-dark-1, 1) 10%, rgba($gray-300, 0.2) 100%);
-    
-    .header__logo {
-      margin-right: $spacing-52;
-      height: 42px;
-    
-      .header__logo__link {
-        height: 42px;
-  
-        .logo1 {
-          height: 42px;
-        }
-      }
-    }
-    
-    .header__menu {
-      @include flexbox(row, nowrap, null, center, $spacing-52);
-      display: flex;
-    }
-  
-    .menu__list {
-      font-size: $text-21;
-    }
-  
-    .list__live {
-      @include flexbox(row, nowrap, center, center, 10px);
-      width: 100px;
-      
-      img {
-        width: 34px;
-      }
-    }
-  
-    .list__paramount{
-      width: 112px;
-  
-      img {
-        width: 112px;
-      }
-    }
-  
-    .header__actions{
-      @include flexbox(row, nowrap, center, center, 40px);
-  
-      .actions__search {
-        width: 40px;
-      }
-  
-      .actions__profile {
-        width: 40px;
-      }
-    }
-  }
-}
+// @include mq(desktop) {
+//   .header {
+//     height: 100px;
+//     padding: 29px 70px;
+//     background: linear-gradient(
+//       to bottom,
+//       rgba($brand-dark-1, 1) 10%,
+//       rgba($gray-300, 0.2) 100%
+//     );
+
+//     .header__logo {
+//       margin-right: $spacing-52;
+//       height: 42px;
+
+//       .header__logo__link {
+//         height: 42px;
+
+//         .logo1 {
+//           height: 42px;
+//         }
+//       }
+//     }
+
+//     .header__menu {
+//       @include flexbox(row, nowrap, null, center, $spacing-52);
+//       display: flex;
+//     }
+
+//     .menu__list {
+//       font-size: $text-21;
+//     }
+
+//     .list__live {
+//       @include flexbox(row, nowrap, center, center, 10px);
+//       width: 100px;
+
+//       img {
+//         width: 34px;
+//       }
+//     }
+
+//     .list__paramount {
+//       width: 112px;
+
+//       img {
+//         width: 112px;
+//       }
+//     }
+
+//     .header__actions {
+//       @include flexbox(row, nowrap, center, center, 40px);
+
+//       .actions__search {
+//         width: 40px;
+//       }
+
+//       .actions__profile {
+//         width: 40px;
+//       }
+//     }
+//   }
+// }
