+++ src/styles/layout/header11.scss
@@ -1,0 +1,172 @@
+@use '@/styles/index' as *;
+
+body {
+  min-width: 320px;
+  width: 100%;
+  background-color: $brand-dark-1;
+  position: relative;
+}
+
+.header {
+  width: 100%;
+  height: 38px;
+  padding: 10px 12px;
+  position: absolute;
+
+  .nav {
+    @include flexbox(row, nowrap, space-between, center, null);
+    color: $gray-200;
+  }
+  
+  .header__logo {
+    display: inline-block;
+    margin-right: $spacing-28;
+    height: 15px;
+  
+    .header__logo__link {
+      display: inline-block;
+      height: 15px;
+
+      .logo1 {
+        height: 15px;
+      }
+    }
+  }
+  
+  .header__menu {
+    @include flexbox(row, nowrap, null, center, $spacing-28);
+    flex-grow: 1;
+    display: none;
+  }
+
+  .menu__list {
+    font-size: $text-12;
+    line-height: $line-160;
+  }
+
+  .list__live {
+    @include flexbox(row, nowrap, center, center, 4px);
+    width: 56px;
+    
+    img {
+      width: 20px;
+    }
+  }
+
+  .list__paramount{
+    width: 65px;
+
+    img {
+      width: 65px;
+    }
+  }
+
+  .header__actions{
+    @include flexbox(row, nowrap, center, center, $spacing-16);
+
+    .actions__search {
+      width: 18px;
+    }
+
+    .actions__profile {
+      width: 18px;
+    }
+  }
+}
+
+@include mq(tablet) {
+  .header {
+    height: 56px;
+    padding: 16px 40px;
+    
+    .header__logo {
+      display: inline-block;
+      margin-right: $spacing-28;
+      height:  24px;
+    
+      .header__logo__link {
+        display: inline-block;
+        height: 24px;
+
+        .logo1 {
+          height: 24px;
+        }
+      }
+    }
+    
+    .header__menu {
+      display: flex;
+    }
+
+    .header__actions{
+      @include flexbox(row, nowrap, center, center, $spacing-24);
+
+      .actions__search {
+        width: 24px;
+      }
+
+      .actions__profile {
+        width: 24px;
+      }
+    }
+  }
+}
+
+@include mq(desktop) {
+  .header {
+    height: 100px;
+    padding: 29px 70px;
+    background: linear-gradient(to bottom, rgba($brand-dark-1, 1) 10%, rgba($gray-300, 0.2) 100%);
+    
+    .header__logo {
+      margin-right: $spacing-52;
+      height: 42px;
+    
+      .header__logo__link {
+        height: 42px;
+  
+        .logo1 {
+          height: 42px;
+        }
+      }
+    }
+    
+    .header__menu {
+      @include flexbox(row, nowrap, null, center, $spacing-52);
+      display: flex;
+    }
+  
+    .menu__list {
+      font-size: $text-21;
+    }
+  
+    .list__live {
+      @include flexbox(row, nowrap, center, center, 10px);
+      width: 100px;
+      
+      img {
+        width: 34px;
+      }
+    }
+  
+    .list__paramount{
+      width: 112px;
+  
+      img {
+        width: 112px;
+      }
+    }
+  
+    .header__actions{
+      @include flexbox(row, nowrap, center, center, 40px);
+  
+      .actions__search {
+        width: 40px;
+      }
+  
+      .actions__profile {
+        width: 40px;
+      }
+    }
+  }
+}
